@@ -1,30 +1,71 @@
 {
-  "name": "lta-express-backend",
-  "version": "0.0.0",
+  "name": "lta-express-backend-api",
+  "version": "1.0.0",
+  "description": "API Backend du projet Latekantique",
+  "main": "./bin/server.ts",
+  "type": "module",
+  "engines": {
+    "node": "20.x"
+  },
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
-  },
+    "start": "node ./bin/server.ts",
+    "startx": "tsx ./bin/server.ts",
+    "startdev": "nodemon ./bin/server.ts",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "migrate:up": "npx sequelize-cli db:migrate",
+    "migrate:undo": "npx sequelize-cli db:migrate:undo"
+  },  
   "dependencies": {
+    "@swc/core": "^1.4.14",
     "@terraformer/arcgis": "^2.1.2",
+    "@types/morgan": "^1.9.9",
     "@types/pug": "^2.0.6",
+    "angular": "^1.8.3",
     "cookie-parser": "~1.4.4",
+    "cors": "^2.8.5",
     "debug": "~2.6.9",
+    "dotenv": "^16.4.5",
+    "emailjs": "^4.0.3",
+    "esbuild-register": "^3.5.0",
     "express": "^4.18.2",
+    "express-bearer-token": "^2.4.0",
+    "express-http-proxy": "^2.0.0",
+    "global": "^4.4.0",
+    "helmet": "^7.1.0",
     "http-errors": "~1.6.3",
+    "jwks-rsa": "^3.1.0",
+    "mailtrap": "^3.2.0",
     "mem": "^9.0.2",
     "morgan": "~1.9.1",
+    "node-fetch": "^3.3.2",
+    "nodemailer": "^6.9.7",
+    "nodemon": "^3.1.0",
     "pug": "^3.0.2",
     "rxjs": "^7.8.1",
+    "sendmail": "^1.6.1",
     "sequelize-auto-migrations": "^1.0.3",
     "sequelize-auto-migrations-ng": "^2.1.1",
+    "stripe": "^14.5.0",
     "terraformer": "^1.0.12",
+    "tsc": "^2.0.4",
+    "tsx": "^4.7.2",
     "validator": "^13.11.0"
   },
   "devDependencies": {
+    "@types/cookie-parser": "^1.4.7",
+    "@types/cors": "^2.8.17",
     "@types/express": "^4.17.17",
+    "@types/express-http-proxy": "^1.6.6",
+    "@types/express-jwt": "^7.4.2",
+    "@types/http-errors": "^2.0.4",
+    "@types/jsonwebtoken": "^9.0.3",
     "@types/node": "^20.4.1",
+    "@types/nodemailer": "^6.4.14",
+    "@types/sendmail": "^1.4.7",
+    "@types/stripe": "^8.0.417",
     "@types/traverse": "^0.6.32",
+    "jsonwebtoken": "^9.0.2",
     "mysql": "^2.18.1",
     "mysql2": "^3.6.1",
     "prisma": "^5.3.0",
@@ -32,7 +73,8 @@
     "sequelize-cli": "^6.6.1",
     "sequelize-cli-typescript": "^3.2.0-c",
     "sequelize-typescript": "^2.1.5",
+    "swc-register": "^1.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
+    "typescript": "^5.4.5"
   }
 }
