@@ -1,78 +1,179 @@
 #!/usr/bin/env node
 
-/**
+/** 
  * Module dependencies.
- */
-var app = require('../app');  
-var debug = require('debug')('lta-express-backend:server');
-var http = require('http');
-const EventEmitter = require('node:events');
+**/   
+import  http from 'http';
+import  debug from 'debug'; //var debug = require('debug')('lta-express-backend:server');
+import  EventEmitter  from 'events';  
+import  AppFactory   from '../src/Application/Factory/AppFactory.ts';    //var App=require('../app');    
+import  App from '../src/Application/app.ts';
+import proxy from 'express-http-proxy';      
+
+import bearerToken from 'express-bearer-token';
+import * as jwt from 'jsonwebtoken';
+import fs from 'fs';
+import AuthorizationController from '../src/Controllers/authorizationController.ts'
 
 
-
-const emitter = new EventEmitter();
+//Définition de Listeners Max (Evite les pb de fuite de mémoire)
+const emitter = new EventEmitter();  
 emitter.setMaxListeners(25);
 
- 
+
+
+
 /**
  * Get port from environment and store in Express.
  */
 var port = normalizePort(process.env.PORT || '3000');
-app.set('port', port);
-
-  
-/**
- * Create HTTP server.
- */
-
-var server = http.createServer(app); 
 
  
+/**
+ * Creation de l'Application
+ */ 
+let app:App | null | undefined;
+
+    let appfactory=AppFactory.initiate();
+     app= appfactory?.getAppInstance(); 
+     app?.getAppRoot();  
+   //  app?.addProperties();  
+     app !== undefined ? console.log('server.ts L32', "app exist"): console.log("server.ts L32 app undefined ");  
+     
+// let app= express();
+// console.log(app);  
+ 
+
+
+/**
+ * Element de sécurisation 
+ **/
+
+//const apiProxy= new  Proxy('localhost:4200',{proxyReqPathResolver: (req:any)=>require('url').parse(req.baseUrl).path});
+   
+ 
+
+//Gestion de proxy (lorsque Express est utilisé derrière un proxy)
+//ou app.set ...  
+app?.getApp().set('trust proxy', function (ip:string){ if (ip === '127.0.0.128' ) return true; //trust ip   // semble ne pas fonctionner  
+                                                       else return false;});      
+  
+//console.log(app?.getApp());
+
+
+
+/** Gestion de requête avec Token
+//app?.getApp().use(bearerToken());
+
+// app?.getApp().use(function (req:Request, res:Response) {
+//   res.send('Token '+req.token);
+// });
+ **/ 
+
+/**
+* Import Middelware
+**/
+import router from '../routes/index.ts';
+import routerPaiement from '../routes/paiement.ts';
+import routerAuthenticate from '../routes/authenticate.ts'  ;
+import { NextFunction, Request, Response } from 'express';
+
+
+
+app?.getApp().use(router);
+app?.getApp().use(routerPaiement); 
+app?.getApp().use(routerAuthenticate);  
+ 
+// Display routes   
+// Combine the routes from all routers   
+const allRoutes = [...routerAuthenticate.stack,...router.stack, ...routerPaiement.stack];
+
+// console.log('All routes in the Express app:');
+// allRoutes.forEach((route) => {
+//   if (route.route) {
+//     console.log(`${route.route.stack[0].method.toUpperCase()} ${route.route.path}`);
+//   }
+// });
+
+ 
+// Middleware function to display route called
+app?.getApp().use(function(req:Request, res:Response, next:NextFunction) {
+    let routeCalled= req.protocol + '://' + req.get('host') + req.originalUrl;
+    console.log("Route called : " + JSON.stringify(routeCalled));    
+    next();    
+});
+
+
+
+// Middleware Authorization
+
+  app?.getApp().use(
+    function(req:Request,res:Response,next:NextFunction){
+    bearerToken( {headerKey: 'Bearer'});
+
+     let authorizationController = new AuthorizationController(req.token);
+
+
+  }); 
+  
 
   
-//console.log(app);
-//let bug=debug('GET /articles') ;  
-//console.log("L33 "+bug); 
+
+
+
+
+
+
+
+
+/**
+ * Create HTTP server.  
+ */ 
+
+//Server without TLS
+var server = http.createServer(app?.getApp()); 
+//console.log(server);  
+
+
+//Server with TLS
+//Configure TLS certificate
+// const serverOptions = {
+// 	// Certificate(s) & Key(s)  when certificate will be configured
+//   	cert: fs.readFileSync(path.join(__dirname, 'Securite/Certificate/host.crt')),
+// 	key: fs.readFileSync(path.join(__dirname, 'Securite/Certificate/privatKey.key')),  
+
+// 	// Optional: TLS Versions
+// 	maxVersion: 'TLSv1.3',
+// 	minVersion: 'TLSv1.2'
+
+// }
+// import https from 'https';
+//const server = https.Server(serverOptions,app?.getApp());
+//Fin Server with certificate  
+
+
+
 
 
 
 /**
  * Listen on provided port, on all network interfaces.
  */
-
-server.listen(port);
+server.listen(port);  
 server.on('error', onError);
 server.on('listening', onListening);
 
 
 
-/**
- * Import Middelware
- */
-let router = require('../routes/index');
 
-  
 
-  
-  
-//console.log(router.stack);
 
-//Affichage des routes
-// app._router.stack.forEach(function(r:any){
-//   //console.log(r);
-//   console.log("Regexp : "+r.regexp); 
-//   console.log("Path : "+r.path); 
-//   if (r.route && r.route.path){
-//     console.log(r.route.path); 
-//   }  
-// })    
-  
 
-/**
+ /**
  * Normalize a port into a number, string, or false.
  */
 
-function normalizePort(val:any) {
+ function normalizePort(val:any) {
   var port = parseInt(val, 10);
 
   if (isNaN(port)) {
@@ -88,10 +189,12 @@ function normalizePort(val:any) {
   return false;
 }
 
+  
+
+
 /**
  * Event listener for HTTP server "error" event.
  */
-
 function onError(error:any) {
   if (error.syscall !== 'listen') {
     throw error;
@@ -116,19 +219,19 @@ function onError(error:any) {
   }
 }
 
+
 /**
  * Event listener for HTTP server "listening" event.
  */
 function onListening() {
   var addr = server.address();
+
   var bind = typeof addr === 'string'
     ? 'pipe ' + addr
-    : 'port ' + addr.port;
+    : 'port ' + addr?.port;
   debug('Listening on ' + bind);
+
 }
 
-//console.log("Index L212") 
-//console.log(router); 
-//console.log(router.stack[0]); 
-
+export default app;   
 
