@@ -1,27 +1,32 @@
 
-import { EvenementType } from '../Types/EvenementType';
+import { EvenementType } from '../Types/EvenementType.ts';
+import { PhotoType } from '../Types/PhotoType.ts';
+import { Photo } from './Photo.ts';
 
 export class Evenement implements EvenementType { 
 
     id: number | null ;  
     title: string| null ;   
-    date : Date [] | null; 
+    date : Date [] | null;   
     description : string| null ; 
-    photo: string| null ;
+    photo: Photo| null ;  
 
     public constructor (
         id: number | null ,  
         title: string| null ,   
         date : Date [] | null, 
         description : string| null , 
-        photo: string| null ,  
+        photo: Photo| null ,  
     ) {
       
         this.id = id ;
-        this.title=title ;   
+        this.title=title ;     
         this.date=date ; 
         this.description=description  ; 
-        this.photo=photo ; 
+        this.photo=photo ;   
 
     }
+    
+   
+   
 }
