@@ -1,40 +1,41 @@
-import { Personne } from "./Interface/Personne";
-import { Address } from './Address' ;
-import { Photo } from './Photo' ;
+import { Personne } from "../Types/Personne.ts";
+import { Address } from './Address.ts' ;
+import { Photo } from './Photo.ts' ;
 
   
 export class Expert implements Personne{
     id : number ; 
     name:string ;
     firstname : string ;
+    lastname: string;
     address:Address ;
     diplomes : string[];
     experiences : string[] ; 
-    email : string ;
+    emailAddress : string ;
     phoneNumber : string ;
-    photo: Photo | null ;
+    photo: Photo | null ;  
     
-   constructor( 
+    constructor( 
                 id : number , 
                 name:string ,
                 firstname : string ,
+                lastname: string,
                 address:Address ,  
                 diplomes:string[],
                 experiences:string[],
-                email : string ,
+                emailAddress : string ,
                 phoneNumber : string ,
                 photo : Photo |null
-){ 
-
+    ){ 
     this.id = id ;
     this.name = name ;
     this.firstname = firstname ;
-    this.address= address ;
+    this.lastname=lastname ;  
+    this.address= address ; 
     this.diplomes = diplomes;
     this.experiences= experiences ; 
-    this.email = email ;
+    this.emailAddress = emailAddress ;
     this.phoneNumber = phoneNumber ;
-    this.photo = photo ;   
-      
+    this.photo = photo ;          
    }
 }
