@@ -1,23 +1,35 @@
-{
-  "development": {
-    "username": "root",
-    "password": null,
-    "database": "database_development",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
+const fs= require('fs');
+
+
+module.exports = {
+  development: {
+    username: "root",
+    password: null,
+    database: "database_development",
+    host: "127.0.0.1",
+    port:3306,
+    dialect: "mysql"
   },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
+  test: {
+    username: process.env.CI_DB_USERNAME,
+    password: process.env.CI_DB_PASSWORD,
+    database: process.env.CI_DB_NAME, // database: "database_test",
+   
+    host: "127.0.0.1",
+    dialect: "mysql"
   },
   "production": {
-    "username": "root",
-    "password": null,
-    "database": "database_production",
-    "host": "127.0.0.1",
-    "dialect": "mysql"
+    username: process.env.PROD_DB_USERNAME,
+    password: process.env.PROD_DB_PASSWORD,
+    database: process.env.PROD_DB_NAME,
+    host: process.env.PROD_DB_HOSTNAME,
+    port: process.env.PROD_DB_PORT,
+    dialect: "mysql",  
+    dialectOptions: {
+      bigNumberStrings: true,
+      ssl: {
+        ca: fs.readFileSync(__dirname + 'Securite/mysql-ca-main.crt'),
+      },
+    },
   }
 }
